feat: create QuantityDisplay component and update MeasurementsView and MedicalActionsView to use it
